Edit and Delete bug fix

+++ src/controller/videoController.js
@@ -47,34 +47,52 @@ export const postUpload = async (req, res, next) => {
 
 export const getEdit = async (req, res, next) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect('/');
   }
   return res.render('edit', { pageTitle: `Edit ${video.title}`, video });
 };
 
 export const postEdit = async (req, res, next) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect('/');
   }
   await Video.findByIdAndUpdate(id, {
     title,
     description,
     hashtags: Video.formatHashtags(hashtags),
   });
+
   return res.redirect(`/videos/${id}`);
 };
 
 export const deleteVideo = async (req, res, next) => {
   const { id } = req.params;
-  console.log(id);
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect('/');
   }
   await Video.findByIdAndDelete(id);
   return res.redirect('/');
